refactor: rename cache.config and remove console.log
