+++ apps/weather/src/infrastructure/metrics/cache.metrics.ts
@@ -44,7 +44,6 @@ export class CacheMetrics implements CacheMetricsInterface {
   }
 
   clearAllMetrics(): void {
-    console.log('Clearing all cache metrics');
     this.cacheHitCounter.reset();
     this.cacheMissCounter.reset();
     this.cacheSize.reset();
